+++ History.md
@@ -1,5 +1,7 @@
 # UNRELEASED
 
+    * requery on ESOCKETTIMEDOUT
+
 # 0.1.0 (2014-10-02)
 
     * initial commit